Split REPL command families into focused modules
